Work on "cranko confirm" to create an `rc` commit
This was more work than anticipated because it exposed the fact that
simple prefix-matching doesn't cut it when you want to determine if a
particular file in a repo is relevant to a particular project. If you've
got nested projects, a not-uncommon situation, you'll need something
more sophisticated. I'm sure my implementation could be cleverer but it
should get us going.