Revise "cranko apply" to pay attention to rc requests
Now using the `ci_info` crate to gather information about the CI
environment. If we seem to be processing the `rc` branch, have `cranko
apply` bump versions based on the serialized request info in the commit
message. Otherwise, bump versions in a "developer" mode. This will
probably need some refinement, but gets at the basic idea.

Here we drop the ReleaseMode enum in favor of VersionBumpScheme, which I
think is a better formalism overall.