Alter the way that changelogs are accumulated.
Now, `cranko stage` will seed the changelog file with its contents from
the most recent release commit on the upstream release branch, or
nothing if no such commit exists. The existing contents of the changelog
in the working tree are discarded. This way, you can use previous
changelog entries for stylistic reference and edit them after-the-fact
if needed.