github: add install-credential-helper utility
This should hopefully ease the process of pushing updates to the source
repository, although all it really does is avoid the need for some
sketchy shell magic if you know what to do. It does also make it so that
the $GITHUB_TOKEN variable doesn't need to be written to disk in the Git
config file (in the naive implementation).