Yet more tuning of the execution-environment code.
My build attempts were failing because `cargo package-released-binaries`
didn't believe that the CI environment was in "release mode", because
that mode could only be activated on updates to the `rc` branch and not
the `master` branch.

Now we have a simpler scheme: CI environments are either in RC or
Release mode, but each of those has a flag for "development mode" or
not. Only updates to the `rc` branch are not in development mode. While
the logic that I had before was going to fail for pull requests and the
like, I think this should DTRT in this circumstances. It also helps us
streamline some of the logic for using the "development mode" RC info
when HEAD is not an RC commit.