cargo: attempt to fix for `cross` with subprojects
Ah, this wasn't working in Tectonic because the switch to subdirectories
for `package-released-binaries` wasn't playing nicely with how I was
trying to invoke `cross` in the subcommand. AFAICT, the more robust
approach is to pass a `--package=` argument to the subcommand, rather
than trying to change its starting directory. So, let's start doing
that.