Start working on splitting up error types
The giant internal error enum is not a good approach, since it exposes
oodles of implementation details. I think that the way to go is to use
specific custom error types when helpful -- implemented with thiserror
-- but in most cases to use the anyhow error type, since we will have
many operations that will have many potential sources of failure.

I thought about wrapping the anyhow Error type so as not to bind
ourselves to it irrevocably, but there are enough useful functions
associated with it that it would be pretty tedious to reimplement them
all. So we'll just expose anyhow.