Fix `release-workflow commit` for new, unreleased projects
We would log release info for every project in the repo. However, this
meant that if you had created a new project, its info would be
propagated to the world even if it wasn't "ready" for release. In
particular, it would get logged in the `release` branch with an age of
0, which would mean that later analysis would think that the project had
been released when it had not.

The solution is to look at the RC submission info, and not log projects
which have never been released and aren't the subject of an RC request
in this run.