Fix apply-versions for some bootstrapped projects.
If you were making the first release of a project with a bootstrap
version, and other projects had been released previously, we would
assign the wrong version because the bootstrap versioning information
would be ignored entirely in favor of the versioning info on the most
recent release commit. The correct solution is to use both sources of
information, but prefer the most recent release commit.