Use a new write_crlf! macro when writing files
It turns out writeln uses \n only, even on Windows. When we rewrite
files it's nicer to preserve line endings so that diffs are smaller
on Windows. Even better would be to actually *preserve* endings based
on file content rather than the current platform, but I don't know
how to do that and don't feel like investigating that right now.