Add the ability to ignore projects
This will be handy for `wwt-webgl-engine`, which has an inner C# project
that we don't release on its own. Besides tidying up output, this also
makes Cranko's change detection more correct.

I *think* that all we need to do here is skip suppressed projects.
Various missing-project error messages could potentially be expanded to
remind the user if a project has been force-ignored. And we could also
report if there are project keys in the config file that seem not to
correspond to anything detected by the software.