zenodo: fix monorepo non-release pathway
I overlooked something. If you were making a release in a monorepo, and
a Zenodo project was *not* being released, the preregister command would
error out. But to be consistent with our pull-request workflow -- and
generally more helpful -- we should succeed without minting new DOIs. In
principle, we should recover the DOIs from the most recent release and
use them, but we don't have great infrastructure to do that. So instead
we just use the existing fake-DOI approach.