npm: apply cargo-like workaround for pre-1.0 caret version ranges
In switching a project to Yarn, I'm running into an issue related to the
fact that semver says that pre-1.0 minor versions are not compatible:
the range `^0.1` does not match `0.2`. I already worked around this with
Cargo, and copy the fix. Except that for Yarn workspace versions, it
looks like I can't apply a `<1` upper bound too.