Work around a bug in which `dynfmt` fails in release mode
See: https://github.com/jan-auer/dynfmt/pull/9

And: https://github.com/jqnatividad/qsv/pull/1467/files

This should go away when a new dynfmt release comes out, or at least
migrate to a scheme where we're not relying on someone else's repo and
branch. But this should be OK for now.